refactor: split view class
